refactor(agent-argocd): collapse to ai-platform-engineering
